requirements.txt
django==3.1.7
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary==2.8.6
drf-yasg==1.20.0
black==20.8b1
pytest-django==4.1.0
factory-boy==3.2.0
flake8==3.9.0
